pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cibil-credit-report-parser"
version = "0.1.0"
description = "Deterministic parser for CIBIL Credit Report PDFs → structured JSON"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "pymupdf4llm>=0.0.10",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]
cloud = [
    "flask>=3.0.0",
    "gunicorn>=22.0.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.19.0",
    "cloudevents>=1.10.0",
]

[project.scripts]
cibil-parser = "cibil_parser.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
